Clean up atomic_pinned.
Switch memory order to use acq_rel when incrementing the counter.

The previous implementation was incorrect. This
should have been more obvious because acquire &
release are only meaningful when done on the same
memory location. We were doing the acquire and
release on different memory locations.

What matters is that when a thread enters the
`is_last_block` branch and loads `temp_accumulator`
that it will see all the updates from all the other threads.
This is guaranteed by acq_rel used to update count_ref.